requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beaupy==3.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
composio_core==0.3.16
composio_langchain==0.3.16
discord.py==2.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
emoji==2.12.1
exceptiongroup==1.2.1
Farama-Notifications==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
inflection==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.82
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.35.7
orjson==3.10.5
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyperclip==1.9.0
Pysher==1.0.8
python-dotenv==1.0.1
python-yakh==0.3.2
pywin32==306
PyYAML==6.0.1
questo==0.2.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
sentry-sdk==2.7.1
simple_parsing==0.1.5
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.2
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
types-requests==2.32.0.20240622
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
zipp==3.19.2
